+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @convergence-rfq/cli
 
+## 4.0.17
+
+### Patch Changes
+
+- add method createEuropeanProgram
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.17
+
 ## 4.0.16
 
 ### Patch Changes