+++ packages/js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @convergence-rfq/sdk
 
+## 4.0.17
+
+### Patch Changes
+
+- add method createEuropeanProgram
+
 ## 4.0.16
 
 ### Patch Changes